More correct promotion for `Rot` constructors
It's common to pass integers to the Rot constructors, but having types
like RotXYZ{Int} are basically useless and confusing because the eltype
as an abstract matrix doesn't match the type of the individual angles.

This is a change to fix the constructors in an attempt to infer a more
correct eltype for the matrix itself.

Notably, this doesn't disentangle the internal representation of the
angles from the logical matrix element type, so this doesn't entirely
fix promotion problems.